updates enumerable extensions to remove any(). Updates equals and gethashcode of state. moves bnf tests
